requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
asynctest==0.13.0
atomicwrites==1.4.1
attrs==23.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blis==0.7.11
bump-pydantic==0.8.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==42.0.8
cymem==2.0.8
datasets==2.19.0
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
environs==11.0.0
exceptiongroup==1.2.1
fastapi==0.110.2
fastapi_cors==0.0.6
filelock==3.13.1
flake8==7.0.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
iniconfig==2.0.0
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.0
langcodes==3.4.0
language_data==1.2.0
libcst==1.3.1
Mako==1.3.5
marisa-trie==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mysqlclient==2.2.4
networkx==3.3
nltk==3.8.1
numpy==1.26.3
openai==1.25.1
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pillow==10.2.0
pluggy==1.5.0
preshed==3.0.9
psycopg2-binary==2.9.9
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.8.0
pydantic-settings==2.2.1
pydantic_core==2.20.0
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.1
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-env==1.1.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.32.2
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
sympy==1.12
tbb==2021.11.0
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tqdm==4.66.2
transformers==4.41.1
typer==0.9.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vaderSentiment==3.3.2
wasabi==1.1.2
weasel==0.3.4
xxhash==3.4.1
yarl==1.9.4
